StyleTransfer/requirements.txt
torch~=1.9.0
torchvision~=0.10.0
matplotlib~=3.4.3
Pillow~=8.3.2